refactor(sql-editor): extract execution + explain hooks
Moves the execute mutation + executeQuery pipeline (destructive-query gating, title gen, connection resolution, potentialIssues state) into useSqlEditorExecution, and the explain mutation + pipeline into useSqlEditorExplain.

Both pipelines accept an already-promoted SafeSqlFragment; acceptUntrustedSql promotion stays in the composition root's run/explain gesture + warning-modal handlers so it remains auditable next to the explicit user action. executeQuery lists its (now all-stable) deps directly instead of suppressing exhaustive-deps. Behavior-preserving. PR 4/6 of the SQLEditor decomposition stack.
